chore: move config logic to internal
