+++ internal/service/domain/auth/auth.go
@@ -2,7 +2,7 @@ package auth
 
 import (
 	"github.com/go-playground/validator/v10"
-	"github.com/green-ecolution/green-ecolution-backend/config"
+	"github.com/green-ecolution/green-ecolution-backend/internal/config"
 	"github.com/green-ecolution/green-ecolution-backend/internal/service"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage"
 )
