+++ internal/storage/auth/storage.go
@@ -1,7 +1,7 @@
 package auth
 
 import (
-	"github.com/green-ecolution/green-ecolution-backend/config"
+	"github.com/green-ecolution/green-ecolution-backend/internal/config"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage"
 	"github.com/green-ecolution/green-ecolution-backend/internal/storage/auth/keycloak"
 )
